feat(multipurpose): add initial version of Python toolbox library
This commit adds the initial version of a Python toolbox library for non-coders and low-coders. The repository comprises categorized functions, such as those related to Selenium automation, and utilities that can be installed via pip or directly cloned from the repository. Users can now import and utilize these functions for easier code reuse and efficiency. Additionally, the creation of a setup script (`setup.py`) has supported the packaging of this library for distribution. This contribution aims to enhance accessibility and ease of use for developers wanting to leverage a collection of reusable functions.

For more comprehensive details, refer to individual file changes within the diff.